+++ CHANGELOG.md
@@ -2,6 +2,10 @@
 
 ## Unreleased
 
+* MSFT_xVMNetworkAdapter:
+  * Added NetworkSettings to be able to statically set IPAddress.
+  * Added option for Vlan tagging. You can now setup a Network Adapeter as an access switch on a specific Vlan.
+
 ## 3.16.0.0
 
 * MSFT_xVMHyperV: